[MOV] product_dimension: split models per file * move fields definition on the top
